split up credentials.ee.test.ts into multiple smaller test files with smaller setups
